requirements.txt
# file: requirements.txt
aiogram==3.6.0
python-dotenv==1.0.1
openai==1.40.0
fastapi==0.115.0
uvicorn[standard]==0.30.6
stripe==7.10.0
pydub==0.25.1
aiohttp~=3.9.0
orjson==3.10.7
apscheduler==3.10.4
